+++ changelog.md
@@ -1,5 +1,10 @@
 # Better Level Browser
 
+## v1.0.2
+
+* Added reload buttons
+* Added background to the levels list
+
 ## v1.0.1
 
 * Fixed lists not working with "better browser" on